[FIX] Don't clean paid invoices when group is changed
- Move clean invoices methods in correct modules
- Remove unused wrappers
